+++ MTP/testing/screens/profile/UICountsPage.swift
@@ -3,7 +3,12 @@
 /// CountsPageVC exposed items
 enum UICountsPage: Exposable {
 
-    case region(Int)
-    case group(Int, Int)
-    case item(Int, Int)
+    /// Counts section header
+    case region(Section)
+    /// Expandable cell under header
+    case group(Section, Row)
+    /// Not expandable cell under header
+    case item(Section, Row)
+    /// Visisted switch in item cell
+    case toggle(Section, Row)
 }